+++ src/eye_extractor/headers.py
@@ -27,6 +27,11 @@ class Headers:
             if initialdict:
                 self.add(initialdict)
 
+    def remove(self, func):
+        """Remove text (e.g., boilerplate) using a function"""
+        self.text = func(self.text)
+        self.data = [{k: func(v) for k, v in d.items()} for d in self.data]
+
     def set_text(self, text):
         self.text = text